requirements.txt
Django==3.2.12
django-environ==0.8.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
django-oauth-toolkit==1.7.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
celery==5.2.3
redis==4.1.4
django-celery-beat==2.2.1
gunicorn==20.1.0